Consolidate the NamedPipeLogger into the FifoLogger
